(refactor) Move repo_display functions to new file
