+++ src/containers/display_helpers.js
@@ -1,0 +1,318 @@
+export default {
+  /**
+   * Create an HTML anchor tag
+   * @param  {String} linkedText - the text you wish to appear
+   * @param  {String} site - the site to link to
+   * @return {String} - the anchor HTML element
+   */
+  makeAnchor: function (linkedText, site) {
+    return `<a href="${site}" target="_blank">${linkedText}</a>`;
+  },
+
+  /**
+   * Generate the x and y-coordinates for each commit. Place them as properties
+   * on the commit.
+   */
+  generateCoordinates: function(sortedCommits, commitObj, allBranches) {
+    /**
+     * Given a range of start and end values, determine if there is any
+     * overlap.
+     * @param  {Object} range1 - the first range, an object with properties
+     *                         start and end
+     * @param  {Object} range2 - the 2nd range, same properties
+     * @return {Boolean} - whether or not they overlap
+     */
+    function checkOverlap(range1, range2) {
+      return (range2.start <= range1.start && range1.start <= range2.end) ||
+        (range2.start <= range1.end && range1.end <= range2.end);
+    }
+
+    /**
+     * Generate an x-value for each commit. Each commit sent in will
+     * have a higher x-value than the one before it. Useful for placing
+     * elements in order. If the next commit will flow off of the screen,
+     * reset the x-value.
+     */
+    function generateX(node) {
+      const nextValue = 40 + numCommits++ * 30;
+      // if(nextValue + 60 > pageWidth) resetXandY();
+      return nextValue;
+    }
+
+    function resetXandY() {
+      numCommits = 1;
+      firstCheckForY += 80;
+    }
+
+    /**
+     * Determine if the y-position we're checking will have overlaps. If so,
+     * put in a different place. Recursively checks the next y-value if the current
+     * one is already taken.
+     * @param  {String} branch - the branch name
+     * @param  {Number} y - the y-value we're checking. Is set automatically,
+     *                    or recursively.
+     * @return {Number} - the y position.
+     */
+    function generateY(branchName, y = firstCheckForY) {
+      //if we're at a new branch we need to jump to another level
+      if(branchName !== lastBranch) {
+        lastBranch = branchName;
+        return generateY(branchName, y + yOffset);
+      }
+
+      let overlap = false;
+
+      const { start: thisBranchStartPoint, end: thisBranchEndPoint } = branchXCoordinates[branchName];
+
+      taken.forEach(set => {
+        if(set.y === y) {
+          if((set.start <= thisBranchStartPoint && thisBranchStartPoint <= set.end) ||
+            (set.start <= thisBranchEndPoint && thisBranchEndPoint <= set.end)) {
+            overlap = true;
+          }
+        }
+      });
+
+      return overlap ? generateY(branchName, y + yOffset) : y;
+    }
+
+    /**
+     * branchXCoordinates will contain the start and end x-values for
+     * each commit.
+     * @type {Object}
+     */
+    const branchXCoordinates = {};
+
+    /**
+     * branchYCoordinates will the y-coordinate of each branch.
+     * @type {Object}
+     */
+    const branchYCoordinates = { master: 360 };
+    let firstCheckForY = 360;
+    let numCommits = 0;
+    const yOffset = 40;
+
+    //Create the x-value for each commit.
+    sortedCommits.forEach(commit => {
+      commit.x = generateX(commit);
+
+      //if it's the first time we're processing a commit from this branch, create an object
+      if(!branchXCoordinates[commit.branch]) {
+        branchXCoordinates[commit.branch] = { start: commit.x };
+      }
+
+      branchXCoordinates[commit.branch].end = commit.x;
+    });
+
+
+    /**
+     * List the positions that are taken. Properties of each object are
+     * a y-value and the range (start and end) of the x-values taken for that
+     * y-value. Initialize with a hard-coded master.
+     * @type {Array}
+     */
+    const taken = [{
+      y: 360,
+      start: branchXCoordinates['master'].start,
+      end: branchXCoordinates['master'].end,
+    }];
+
+    /**
+     * A variable to store the last branch location. Every commit, when being
+     * placed, will be compared to the last commit to see if the branch was
+     * different. If it was different, the new branch will be placed lower.
+     */
+    let lastBranch;
+
+    //get the y-coordinates for each branch
+    Object.keys(allBranches).forEach(branch => {
+      if(!branchXCoordinates[branch] || branch === 'master') return;
+
+      const { start, end } = branchXCoordinates[branch];
+      const yCoordinate = generateY(branch);
+      lastBranch = branch;
+      branchYCoordinates[branch] = yCoordinate;
+      taken.push({ start, end, y: yCoordinate });
+    });
+
+    // if there are 2 branches connected and on the same line, move one
+    let changed = false;
+    do{
+      changed = false;
+      sortedCommits.forEach(commit => {
+        commit.children.forEach(child => {
+          const childObj = commitObj[child];
+          if(childObj && commit.branch !== childObj.branch) {
+            if(branchYCoordinates[commit.branch] === branchYCoordinates[childObj.branch]) {
+              branchYCoordinates[childObj.branch] += yOffset;
+              changed = true;
+            }
+          }
+        });
+      });
+    } while(changed);
+
+    //If there are 2 branches overlapping, move one
+    const branchNames = Object.keys(branchXCoordinates);
+    let altered = false;
+    do{
+      altered = false;
+      branchNames.forEach(thisBranch => {
+        const thisBranchSet = branchXCoordinates[thisBranch];
+        branchNames.forEach(branchToCheck => {
+          //make sure it's not the same branch
+          if(thisBranch !== branchToCheck){
+            //make sure they have the same y-coordinate
+            if(branchYCoordinates[thisBranch] === branchYCoordinates[branchToCheck]){
+              const branchToCheckSet = branchXCoordinates[branchToCheck];
+
+              //Make sure they overlap somewhere along their x-coordinates
+              if(thisBranch !== branchToCheck && checkOverlap(thisBranchSet, branchToCheckSet)) {
+                branchYCoordinates[branchToCheck] += yOffset;
+                altered = true;
+              }
+            }
+          }
+        });
+      });
+    } while(altered);
+
+    //map the branchYCoordinates values over to their commits
+    sortedCommits.forEach(commit => commit.y = branchYCoordinates[commit.branch]);
+  },
+
+  //https://bl.ocks.org/mbostock/6123708
+  zoomed: function() {
+    const { translate, scale } = d3.event;
+    svg.selectAll('g')
+      .attr("transform", "translate(" + translate + ")scale(" + scale + ")");
+    svg.selectAll('line')
+      .attr("transform", "translate(" + translate + ")scale(" + d3.event.scale + ")");
+
+    //To make the tip essentially disappear from the page we remove its HTML.
+    //It is still present on the page, but now consists of a tiny invisible square.
+    d3.selectAll('.d3-tip')
+      .style('opacity', 0)
+      .html('');
+  },
+
+  //give each branch a different color property
+  addColors: function(branches) {
+    const colors = [
+      '#FA8072',  //salmon
+      "#7D3C98",  //purple
+      "#2471A3",  //blue
+      "#F1C40F",  //yellow
+      '#E67E22',  //orange
+      "#A93226",  //red
+      "#17A589",  //aqua
+      '#839192',  //grey
+      '#000000',  //black
+    ];
+
+    let i = 0;
+    for(let branch in branches) {
+      branches[branch].color = colors[i++ % colors.length];
+    }
+  },
+
+  //Create an object with some statistics for the commits passed in
+  analyzeRepo: function(commits, branchLookup) {
+    function countCommitsPerAuthor(author) {
+      return commits.reduce((authorCount, commit) => {
+        if(commit.author.login === author) authorCount++;
+        return authorCount;
+      }, 0);
+    }
+
+    function countCommitsPerBranch(branch) {
+      return commits.reduce((branchCount, commit) => {
+        if(commit.branch === branch) branchCount++;
+        return branchCount;
+      }, 0);
+    }
+
+    const stats = {
+      branches: {},
+      contributors: {},
+    };
+
+    const { branches, contributors } = stats;
+
+    for(let branch in branchLookup) {
+      branches[branch] = branches[branch] || countCommitsPerBranch(branch);
+    }
+
+    commits.forEach((commit) => {
+      const author = commit.author.login;
+      contributors[author] = contributors[author] || countCommitsPerAuthor(author);
+    });
+
+    return stats;
+  },
+
+  //////////////////////////////////////////////////////////////////
+  // NOT DRY. COPIED FROM ORIGINAL LINE RENDERING IN repo_display //
+  //////////////////////////////////////////////////////////////////
+  flipXY: function() {
+    d3.selectAll('circle')
+      .each(function(node) {
+        const x = node.y;
+        const y = node.x;
+        d3.select(this)
+          .attr('cx', x)
+          .attr('cy', y);
+      });
+
+    d3.selectAll('.line')
+      .remove();
+
+    d3commits.forEach(commit => {
+      commit.children.forEach(child => {
+        let childObj = githubTranslator.getCommit(child);
+
+        //this is where the magic happens. Just flip x and y.
+        const curveData = [
+          { x:commit.y, y:commit.x },
+          { x:childObj.y,  y:childObj.x },
+        ];
+
+        const edge = d3.select("svg").append('g');
+        const diagonal = d3.svg.diagonal()
+          .source(function(d) {return {"x":d[0].y, "y":d[0].x}; })            
+          .target(function(d) {return {"x":d[1].y, "y":d[1].x}; })
+          .projection(function(d) { return [d.y, d.x]; });
+           
+        d3.select("g")
+            .datum(curveData)
+          .append("path")
+            .attr("class", "line")
+            .attr("d", diagonal)
+            .attr("stroke-width", 1)
+          .attr('stroke', branchLookup[commit.branch].color)
+          .attr('fill', 'none');
+      });
+    });
+  },
+
+  //Confusing indentation due to the string literal:
+  //-----------------------------------------------------------------------------
+  showToolTip: function(commit) {
+    const { branch, sha, html_url: url, author: { login: authorName } } = commit;
+    const repoName = url.match(/\/\/[\w\.]*\/[\w\.]*\/(\w*)\//);
+
+    //the ternary operator below: if the branch name is not fake (e.g. master, dev, etc.)
+    //then make it a hyperlink; otherwise, don't display branch name
+    const branchLink = `https://github.com/${authorName}/${repoName[1]}/commits/${branch}`;
+
+    const tooltipContent =
+`${originalBranches.includes(branch) ? 'Branch: ' + makeAnchor(branch, branchLink) + '\n' : '' }SHA:     ${makeAnchor(sha.slice(0, 9) + '...', url)}
+Author:  ${authorName}
+
+Message: ${commit.commit.message}`;
+
+    infoTip.html(`<pre>${tooltipContent}</pre>`);
+    infoTip.show();
+  },
+  //-----------------------------------------------------------------------------
+};
